results/final_count/run-20230926-130438_LLMatic-zero_shot_donnemartin-system-design-oop-whatsapp-donnemartin-system-design-oop-url_shortener-donnemartin-system-design-oop-dropbox-donnemartin-system-design-oop-twitter_0_10-runs_log/dropbox/0/LLMatic/tests/test_user_login.py
import pytest
from services.user_service import UserService

user_service = UserService()

def test_authenticate_user():
	user_service.register_user('Test User', 'test@example.com', 'password')
	user = user_service.authenticate_user('test@example.com', 'password')
	assert user is not None
	assert user['email'] == 'test@example.com'
	assert user['password'] == 'password'

	invalid_user = user_service.authenticate_user('invalid@example.com', 'password')
	assert invalid_user is None

	invalid_password = user_service.authenticate_user('test@example.com', 'wrongpassword')
	assert invalid_password is None
